+++ qupath-gui-fx/src/main/java/qupath/lib/gui/panes/ImageDetailsPane.java
@@ -26,7 +26,6 @@ package qupath.lib.gui.panes;
 import java.awt.image.BufferedImage;
 import java.beans.PropertyChangeEvent;
 import java.beans.PropertyChangeListener;
-import java.io.File;
 import java.io.IOException;
 import java.net.URI;
 import java.net.URLDecoder;
@@ -35,6 +34,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Locale;
@@ -43,8 +43,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.stream.IntStream;
 
-import javax.imageio.ImageIO;
-
+import javafx.scene.control.TitledPane;
 import org.controlsfx.control.MasterDetailPane;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -55,7 +54,6 @@ import javafx.beans.property.ReadOnlyObjectWrapper;
 import javafx.beans.property.ReadOnlyStringWrapper;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
-import javafx.embed.swing.SwingFXUtils;
 import javafx.event.ActionEvent;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
@@ -63,14 +61,10 @@ import javafx.geometry.Side;
 import javafx.scene.control.Label;
 import javafx.scene.Group;
 import javafx.scene.Node;
-import javafx.scene.Scene;
 import javafx.scene.control.ButtonType;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.ContentDisplay;
 import javafx.scene.control.ListView;
-import javafx.scene.control.Menu;
-import javafx.scene.control.MenuBar;
-import javafx.scene.control.MenuItem;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
@@ -78,17 +72,10 @@ import javafx.scene.control.ToggleButton;
 import javafx.scene.control.ToggleGroup;
 import javafx.scene.control.Tooltip;
 import javafx.scene.effect.DropShadow;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.input.Clipboard;
-import javafx.scene.input.ClipboardContent;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyCodeCombination;
-import javafx.scene.input.KeyCombination;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.input.MouseEvent;
-import javafx.scene.layout.Background;
-import javafx.scene.layout.BackgroundFill;
 import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.ColumnConstraints;
 import javafx.scene.layout.GridPane;
@@ -99,25 +86,18 @@ import javafx.scene.paint.Color;
 import javafx.scene.shape.Ellipse;
 import javafx.scene.shape.Rectangle;
 import javafx.scene.text.TextAlignment;
-import javafx.stage.Modality;
-import javafx.stage.Stage;
 import qupath.fx.utils.FXUtils;
-import qupath.fx.dialogs.FileChoosers;
-import qupath.lib.awt.common.BufferedImageTools;
 import qupath.lib.color.ColorDeconvolutionHelper;
 import qupath.lib.color.ColorDeconvolutionStains;
 import qupath.lib.color.StainVector;
 import qupath.lib.common.ColorTools;
 import qupath.lib.common.GeneralTools;
-import qupath.lib.display.ChannelDisplayInfo;
-import qupath.lib.display.ImageDisplay;
 import qupath.lib.gui.QuPathGUI;
 import qupath.fx.dialogs.Dialogs;
 import qupath.lib.gui.dialogs.ParameterPanelFX;
 import qupath.lib.gui.prefs.PathPrefs;
 import qupath.lib.gui.prefs.PathPrefs.ImageTypeSetting;
 import qupath.fx.utils.GridPaneUtils;
-import qupath.lib.gui.prefs.SystemMenuBar;
 import qupath.lib.gui.tools.GuiTools;
 import qupath.lib.images.ImageData;
 import qupath.lib.images.ImageData.ImageType;
@@ -125,7 +105,6 @@ import qupath.lib.images.servers.ImageServer;
 import qupath.lib.images.servers.ImageServerMetadata;
 import qupath.lib.images.servers.PixelCalibration;
 import qupath.lib.images.servers.ServerTools;
-import qupath.lib.images.servers.WrappedBufferedImageServer;
 import qupath.lib.plugins.parameters.ParameterList;
 import qupath.lib.plugins.workflow.DefaultScriptableWorkflowStep;
 import qupath.lib.plugins.workflow.WorkflowStep;
@@ -152,6 +131,8 @@ public class ImageDetailsPane implements ChangeListener<ImageData<BufferedImage>
 
 	private TableView<ImageDetailRow> table = new TableView<>();
 	private ListView<String> listAssociatedImages = new ListView<>();
+
+	private Map<String, SimpleImageViewer> associatedImageViewers = new HashMap<>();
 
 	private enum ImageDetailRow {
 		NAME, URI, PIXEL_TYPE, MAGNIFICATION, WIDTH, HEIGHT, DIMENSIONS,
@@ -204,7 +185,11 @@ public class ImageDetailsPane implements ChangeListener<ImageData<BufferedImage>
 
 		MasterDetailPane mdPane = new MasterDetailPane(Side.BOTTOM);
 		mdPane.setMasterNode(new StackPane(table));
-		mdPane.setDetailNode(new StackPane(listAssociatedImages));
+		var titlePaneAssociated = new TitledPane("Associated images", listAssociatedImages);
+		titlePaneAssociated.setCollapsible(false);
+		listAssociatedImages.setTooltip(new Tooltip(
+				"Extra images associated with the current image, e.g. a label or thumbnail"));
+		mdPane.setDetailNode(titlePaneAssociated);
 		mdPane.showDetailNodeProperty().bind(
 				Bindings.createBooleanBinding(() -> !listAssociatedImages.getItems().isEmpty(),
 						listAssociatedImages.getItems()));
@@ -217,83 +202,27 @@ public class ImageDetailsPane implements ChangeListener<ImageData<BufferedImage>
 		if (event.getClickCount() < 2 || listAssociatedImages.getSelectionModel().getSelectedItem() == null)
 			return;
 		String name = listAssociatedImages.getSelectionModel().getSelectedItem();
-
-		Stage dialog = new Stage();
-		dialog.setTitle(name);
-		dialog.initModality(Modality.NONE);
-		var qupath = QuPathGUI.getInstance();
-		if (qupath != null)
-			dialog.initOwner(qupath.getStage());
-
-		// Create menubar
-		MenuBar menubar = new MenuBar();
-		Menu menuFile = new Menu("File");
-		MenuItem miClose = new MenuItem("Close");
-		miClose.setAccelerator(new KeyCodeCombination(KeyCode.W, KeyCombination.SHORTCUT_DOWN));
-		miClose.setOnAction(e -> dialog.close());
-		MenuItem miSave = new MenuItem("Save as PNG");
-		miSave.setAccelerator(new KeyCodeCombination(KeyCode.S, KeyCombination.SHORTCUT_DOWN));
-		miSave.setOnAction(e -> {
-			BufferedImage img = imageData.getServer().getAssociatedImage(name);
-			File fileOutput = FileChoosers.
-					promptToSaveFile(dialog, "Save image",
-							name == null ? null : new File(name),
-							FileChoosers.createExtensionFilter("PNG", ".png"));
-			if (fileOutput != null) {
-				try {
-					ImageIO.write(img, "PNG", fileOutput);
-				} catch (Exception e1) {
-					Dialogs.showErrorMessage("Save image", "Error saving " + fileOutput.getName() + "\n" + e1.getLocalizedMessage());
-				}
-			}
-		});
-		menuFile.getItems().addAll(miSave, miClose);
-
-		Menu menuEdit = new Menu("Edit");
-		MenuItem miCopy = new MenuItem("Copy");
-		miCopy.setAccelerator(new KeyCodeCombination(KeyCode.C, KeyCombination.SHORTCUT_DOWN));
-		miCopy.setOnAction(e -> {
-			Image img = SwingFXUtils.toFXImage(imageData.getServer().getAssociatedImage(name), null);
-			ClipboardContent content = new ClipboardContent();
-			content.putImage(img);
-			Clipboard.getSystemClipboard().setContent(content);
-		});
-
-
-		menuEdit.getItems().add(miCopy);
-
-		menubar.getMenus().addAll(menuFile, menuEdit);
-
-		// Create image view
-		var imgBuffered = imageData.getServer().getAssociatedImage(name);
-		if (!BufferedImageTools.is8bitColorType(imgBuffered.getType()) && imgBuffered.getType() != BufferedImage.TYPE_BYTE_GRAY) {
-			// By wrapping the thumbnail, we avoid slow z-stack/time series requests & determine brightness & contrast just from one plane
-			var wrappedServer = new WrappedBufferedImageServer("Dummy", imgBuffered);
-			var imageDisplay = new ImageDisplay(new ImageData<>(wrappedServer));
-			for (ChannelDisplayInfo info : imageDisplay.selectedChannels()) {
-				imageDisplay.autoSetDisplayRange(info);
-			}
-			imgBuffered = imageDisplay.applyTransforms(imgBuffered, null);
+		var simpleViewer = associatedImageViewers.get(name);
+		if (simpleViewer == null) {
+			simpleViewer = new SimpleImageViewer();
+			var img = imageData.getServer().getAssociatedImage(name);
+			simpleViewer.updateImage(name, img);
+			var stage = simpleViewer.getStage();
+			var owner = FXUtils.getWindow(getPane());
+			stage.initOwner(owner);
+			stage.setOnCloseRequest(e -> {
+				associatedImageViewers.remove(name);
+				stage.close();
+				e.consume();
+			});
+			// Show with constrained size (in case we have a large image)
+			GuiTools.showWithScreenSizeConstraints(stage, 0.8);
+			associatedImageViewers.put(name, simpleViewer);
+		} else {
+			simpleViewer.getStage().show();
+			simpleViewer.getStage().toFront();
 		}
-		Image img = SwingFXUtils.toFXImage(imgBuffered, null);
-		ImageView imageView = new ImageView(img);
-		BorderPane pane = new BorderPane();
-		imageView.fitWidthProperty().bind(pane.widthProperty());
-		imageView.fitHeightProperty().bind(pane.heightProperty());
-		imageView.setPreserveRatio(true);
-		pane.setCenter(imageView);
-		pane.setTop(menubar);
-		Scene scene = new Scene(pane);
-		pane.prefWidthProperty().bind(scene.widthProperty());
-		pane.prefHeightProperty().bind(scene.heightProperty());
-		SystemMenuBar.manageChildMenuBar(menubar);
-		//				menubar.setUseSystemMenuBar(true);
-		pane.setBackground(new Background(new BackgroundFill(Color.BLACK, null, null)));
-
-		dialog.setScene(scene);
-		dialog.show();
 	}
-	
 
 
 	private static boolean hasOriginalMetadata(ImageServer<BufferedImage> server) {
@@ -728,6 +657,17 @@ public class ImageDetailsPane implements ChangeListener<ImageData<BufferedImage>
 				listAssociatedImages.getItems().clear();
 			else
 				listAssociatedImages.getItems().setAll(server.getAssociatedImageList());
+		}
+
+		// Check if we're showing associated images
+		for (var entry : associatedImageViewers.entrySet()) {
+			var name = entry.getKey();
+			var simpleViewer = entry.getValue();
+			logger.trace("Updating associated image viewer for {}", name);
+			if (server == null || !server.getAssociatedImageList().contains(name))
+				simpleViewer.updateImage(name, (BufferedImage)null); // Hack to retain the title, without an image
+			else
+				simpleViewer.updateImage(name, server.getAssociatedImage(name));
 		}
 	}